refactor : 'LineService' 리팩터링
